Fully commented
Updated and fully commented version of code used for Macrobdella manuscript. This is the complete code and is separated by figure/table number for which the code is written. Due to the modularity of the code, variables are defined multiple times but commented out for all but the first time. If using the code modularly, you may need to uncomment a variable definition to make it work.